Update templates.xml
The removed line causes a lua error on every arena entry. It refers to a texture that no longer exists, and was removed with the UI update. 

I considered replacing it with the new equivalent texture, but after looking through the code, it doesn't appear we use it anywhere. 

This PR removes an outdated texture inheritance, and fixes a common error cause.