MoP Support + Others (#118)
* Fixed aura mastery effect dissipation on Classic

* Fixed typo

* Fix for Classic

* Minor rework for UnpackAuraData

Made it UnpackAuraData local and optimized it a bit (avoiding unpack call). I'm not sure why we need this function at all, can anyone explain?

* Spaces to tabs in interrupts.lua

Not sure if it should be spaces or tabs, but at least now the file is internally consistent.

* Fixed small mistake in interrupts module

* Fixed auras for Classic

* Precognito Absorb Support

- Added Precognito support for Classic (absorb tracker) similar to sArena

* Added WoD Deterministic DR Support

- Prepped some WoD support (IS_PRE_WOD) for DRTracker (in WoD DRs are deterministic instead of random)

* Cleanup

* TBCC Preparation

Will need more stuff to actually work (like adding the correct spellIDs to Data-BCC, correct spell IDs in LibCooldownTracker for TBCC etc).

* Added full MoP support

Should work now.

* Added GUI support for custom PvP trinket icons

Requires https://github.com/vendethiel/LibCooldownTracker-10/pull/43 to be merged and utilized in GladiusEx for it to work.

* Changed from WotLK to _wotlk